+++ src/Models/Forms/ConsoleSignupForm.php
@@ -22,6 +22,7 @@ class ConsoleSignupForm extends AbstractSignupForm
                 ['password'],
                 'string',
                 'min' => $this->user->passwordMinLength,
+                'max' => $this->user->passwordMaxLength,
             ],
         ];
     }